Queen of hearts adjustments

+++ app/Game/NpcActions/QueenOfHeartsActions/Services/ReRollEnchantmentService.php
@@ -190,26 +190,15 @@ class ReRollEnchantmentService {
                 }
             }
 
-            if ($deletedOne && $deletedTwo && $slot->item->appliedHolyStacks->isEmpty()) {
+            if ($deletedOne && $deletedTwo && $slot->item->appliedHolyStacks->isEmpty() && $slot->item->socket_count === 0) {
                 $slot->delete();
                 $duplicateUnique->delete();
 
                 $deletedAll = true;
             } else {
 
-                foreach ($slot->item->appliedHolyStacks as $stack) {
-                    $duplicateUnique->appliedHolyStacks()->create([
-                        'item_id'                  => $duplicateUnique->id,
-                        'devouring_darkness_bonus' => $stack->devouring_darkness_bonus,
-                        'stat_increase_bonus'      => $stack->stat_increase_bonus,
-                    ]);
-
-                    $duplicateUnique = $duplicateUnique->refresh();
-                }
-
                 if ($deletedOne && $deletedTwo) {
                     $duplicateUnique->update([
-                        'market_sellable' => false,
                         'is_mythic' => false,
                     ]);
